requirements.txt
gunicorn
psycopg2-binary
Django
django-admin-interface
django-colorfield
django-environ
django-jalali
django-simple-captcha
django-rest-knox
djangorestframework
django-auditlog==2.3.0

Markdown==3.5.1
cryptography==41.0.5
Pillow==10.1.0
pytz==2023.3.post1
requests==2.31.0
googletrans==4.0.0-rc1


